requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-active-link==0.1.8
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-currentuser==0.5.3
django-decorator-include==3.0
django-filter==22.1
django-import-export==2.8.0
django-select2==7.10.0
django-sql-explorer==2.4.1
django-tables2==2.4.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pycodestyle==2.8.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
django-resized==1.0.2
pandas==2.0.3
